trading_bot/requirements.txt
fastapi==0.115.8
Flask==3.0.3
numpy==1.23.5
pandas==1.5.3
python-dotenv==1.0.1
Requests==2.32.3
textblob==0.18.0
torch==2.2.2
